Payment for SubscriptionPackage and LeaseAgreement aligned using polymorphism, will align the filament page later

+++ database/migrations/2025_06_16_161409_create_payments_table.php
@@ -25,7 +25,7 @@ return new class extends Migration
                 ->constrained()
                 ->nullOnDelete();
             $table->date('payment_date');
-            $table->decimal('payment_amount', 10);
+            $table->integer('payment_amount');
             $table->timestamps();
         });
     }
